riskengine/requirements.txt
altair==5.1.2
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
clarabel==0.6.0
click==8.1.7
cvxpy==1.4.1
ecos==2.0.12
Flask==3.0.0
frozendict==2.3.8
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
gunicorn==21.2.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multitasking==0.0.11
numpy==1.26.1
osqp==0.6.3
packaging==23.2
pandas==2.1.2
peewee==3.17.0
Pillow==10.1.0
protobuf==4.25.0
pyarrow==14.0.0
pybind11==2.11.1
pydeck==0.8.1b0
Pygments==2.16.1
pyportfolioopt==1.5.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
qdldl==0.1.7.post0
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
scipy==1.11.3
scs==3.2.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.28.1
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.15
validators==0.22.0
webencodings==0.5.1
Werkzeug==3.0.1
yfinance==0.2.31
zipp==3.17.0
